+++ resources/lang/en/logs-admin.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'switch-logfile'    => 'Switch logfile',
+    'change'            => 'Change',
+];